+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased] - TBD
 
+## [3.0.0a3] - 2024-06-13
+
 ### Added
 
 * Add base support for the Aggregation extension [#684](https://github.com/stac-utils/stac-fastapi/pull/684)
@@ -396,7 +398,8 @@
 
 * First PyPi release!
 
-[Unreleased]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.0a2..main>
+[Unreleased]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.0a3..main>
+[3.0.0a3]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.0a2..3.0.0a3>
 [3.0.0a2]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.0a1..3.0.0a2>
 [3.0.0a1]: <https://github.com/stac-utils/stac-fastapi/compare/3.0.0a0..3.0.0a1>
 [3.0.0a0]: <https://github.com/stac-utils/stac-fastapi/compare/2.5.5.post1..3.0.0a0>